backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-binance==1.0.19
pandas==2.1.4
numpy==1.24.3
redis==5.0.1
requests==2.31.0
pydantic==2.5.0
tenacity==8.2.3
numba==0.58.1
structlog==23.2.0
python-multipart==0.0.6
scikit-learn==1.3.2
aiohttp==3.9.1
httpx[http2]==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
joblib==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tensorflow==2.15.0
pytorch-lightning==2.1.2
torch==2.1.2
ta-lib==0.4.28
yfinance==0.2.28
ccxt==4.1.77
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2
scipy==1.11.4
statsmodels==0.14.1
arch==6.2.0
quantlib==1.32
asyncio-mqtt==0.16.1
alpaca-trade-api==3.1.1
websocket-client==1.6.4
transformers==4.36.2
huggingface-hub==0.19.4
datasets==2.15.0
accelerate==0.25.0
sentencepiece==0.1.99
tokenizers==0.15.0
python-dotenv==1.0.0